Fix imports and rename common/calendar/import to common/calendar/gui
Lazyload EventBannerImpl.ts to allow EventBanner.ts to stay with MailView
Adds a new directory mail-app/gui/date to date-gui chunk allowing calendar
related ui to stay inside mail-app folder structure.
