+++ buildSrc/RollupConfig.js
@@ -183,7 +183,8 @@ export function getChunkName(moduleId, { getModuleInfo }) {
 		isIn("src/common/ratings") ||
 		isIn("src/calendar-app/calendar/model") ||
 		isIn("src/calendar-app/calendar/gui") ||
-		isIn("src/common/calendar/import")
+		isIn("src/common/calendar/gui") ||
+		isIn("src/mail-app/gui/date")
 	) {
 		// this contains code that is important to the calendar view but might be used by other parts of the app on the main thread
 		// like time-based input components and formatting code.
