+++ src/calendar-app/calendar/gui/EditCalendarDialog.ts
@@ -8,7 +8,7 @@ import type { TranslationKeyType } from "../../../common/misc/TranslationKey.js"
 import { deepEqual, isNotNull } from "@tutao/tutanota-utils"
 import { AlarmInterval, CalendarType, isExternalCalendarType, isNormalCalendarType } from "../../../common/calendar/date/CalendarUtils.js"
 import { RemindersEditor } from "./RemindersEditor.js"
-import { checkURLString, isIcal } from "../../../common/calendar/import/ImportExportUtils.js"
+import { checkURLString, isIcal } from "../../../common/calendar/gui/ImportExportUtils.js"
 import { locator } from "../../../common/api/main/CommonLocator.js"
 import type { CalendarModel } from "../model/CalendarModel.js"
 import { DEFAULT_ERROR } from "../../../common/api/common/TutanotaConstants.js"
