+++ src/calendar-app/calendar/view/CalendarInvites.ts
@@ -1,4 +1,4 @@
-import { parseCalendarFile } from "../../../common/calendar/import/CalendarImporter.js"
+import { parseCalendarFile } from "../../../common/calendar/gui/CalendarImporter.js"
 import type { CalendarEvent, CalendarEventAttendee, File as TutanotaFile, Mail, MailboxProperties } from "../../../common/api/entities/tutanota/TypeRefs.js"
 import { locator } from "../../../common/api/main/CommonLocator.js"
 import { CalendarAttendeeStatus, CalendarMethod, ConversationType, FeatureType, getAsEnumValue } from "../../../common/api/common/TutanotaConstants.js"
@@ -27,13 +27,13 @@ import { newPromise } from "@tutao/tutanota-utils/dist/Utils"
 // not picking the status directly from CalendarEventAttendee because it's a NumberString
 export type Guest = Recipient & { status: CalendarAttendeeStatus }
 
-export type ParsedIcalFileContent =
-	| {
-			method: CalendarMethod
-			events: Array<CalendarEvent>
-			uid: string
-	  }
-	| None
+export interface ParsedIcalFileContentData {
+	method: CalendarMethod
+	events: Array<CalendarEvent>
+	uid: string
+}
+
+export type ParsedIcalFileContent = ParsedIcalFileContentData | None
 
 async function getParsedEvent(fileData: DataFile): Promise<ParsedIcalFileContent> {
 	try {
