+++ src/common/misc/DeviceConfig.ts
@@ -9,7 +9,7 @@ import { client } from "./ClientDetector"
 import { NewsItemStorage } from "./news/NewsModel.js"
 import { CredentialsInfo } from "../native/common/generatedipc/CredentialsInfo.js"
 import { CalendarViewType } from "../api/common/utils/CommonCalendarUtils.js"
-import { SyncStatus } from "../calendar/import/ImportExportUtils.js"
+import { SyncStatus } from "../calendar/gui/ImportExportUtils.js"
 import Stream from "mithril/stream"
 import stream from "mithril/stream"
 import type { GroupSettings } from "../api/entities/tutanota/TypeRefs.js"
